requirements.txt
Flask
Flask-RESTful
Pillow
numpy
tflite-runtime
gunicorn